Extract a `Har` namespace
We were starting to have a lot of classes prefixed with `Har` so this
extracts a separate namespace for them. It also renames some of the
generic classes as serializers to better represent their role now that
some of the more model-like responsibilities have been extracted to
standalone classes like `HttpRequest`.

As part of making sure all the specs worked, I ran a lot of them
independently and found that they didn't have all the `require`s they
needed to do so. Ideally, a spec file would only load the system under
test and any testing dependencies. Everything else needed would be
required by the lib file.

I removed the `require "readme/metrics` from `spec_helper.rb` since this
doesn't need to be required in ever spec file.
