+++ packages/ruby/lib/readme/har/collection.rb
@@ -1,0 +1,24 @@
+module Readme
+  module Har
+    class Collection
+      def initialize(filter, hash)
+        @filter = filter
+        @hash = hash
+      end
+
+      def to_h
+        filtered_hash
+      end
+
+      def to_a
+        filtered_hash.map { |name, value| {name: name, value: value} }
+      end
+
+      private
+
+      def filtered_hash
+        @filter.filter(@hash)
+      end
+    end
+  end
+end
