+++ packages/ruby/spec/http_request_spec.rb
@@ -1,4 +1,3 @@
-require "spec_helper"
 require "http_request"
 
 RSpec.describe HttpRequest do
