+++ packages/ruby/spec/readme/har/collection_spec.rb
@@ -1,12 +1,11 @@
-require "readme/har_collection"
-require "readme/filter"
+require "readme/har/collection"
 
-RSpec.describe Readme::HarCollection do
+RSpec.describe Readme::Har::Collection do
   describe "#to_h" do
     it "returns a hash with filtering applied" do
       hash = {keep: "keep", reject: "reject"}
       filter = double(filter: {keep: "kept"})
-      collection = Readme::HarCollection.new(filter, hash)
+      collection = Readme::Har::Collection.new(filter, hash)
 
       expect(collection.to_h).to eq({keep: "kept"})
     end
@@ -16,7 +15,7 @@ RSpec.describe Readme::HarCollection do
     it "returns an array with filtering applied" do
       hash = {keep: "keep", reject: "reject"}
       filter = double(filter: {keep: "kept"})
-      collection = Readme::HarCollection.new(filter, hash)
+      collection = Readme::Har::Collection.new(filter, hash)
 
       expect(collection.to_a).to eq([{name: :keep, value: "kept"}])
     end
