+++ packages/ruby/spec/readme/metrics_spec.rb
@@ -1,3 +1,4 @@
+require "readme/metrics"
 require "rack/test"
 require "webmock/rspec"
