+++ packages/ruby/spec/spec_helper.rb
@@ -1,5 +1,5 @@
 require "bundler/setup"
-require "readme/metrics"
+
 Dir["#{Dir.pwd}/spec/support/**/*.rb"].sort.each { |f| require f }
 
 RSpec.configure do |config|
